Test: add test for changing entry dates
