+++ lib/widgets/journal/entry_details/duration_widget.dart
