+++ test/widgets/journal/entry_details/entry_detail_header_test.dart
@@ -9,6 +9,7 @@ import 'package:lotti/routes/router.gr.dart';
 import 'package:lotti/services/tags_service.dart';
 import 'package:lotti/themes/themes_service.dart';
 import 'package:lotti/widgets/journal/entry_details/entry_detail_header.dart';
+import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:mocktail/mocktail.dart';
 
@@ -30,6 +31,8 @@ void main() {
         ..registerSingleton<JournalDb>(JournalDb(inMemoryDatabase: true))
         ..registerSingleton<TagsService>(TagsService())
         ..registerSingleton<AppRouter>(mockAppRouter);
+
+      when(mockAppRouter.pop).thenAnswer((_) async => true);
 
       when(() => entryCubit.state).thenAnswer(
         (_) => EntryState.dirty(
@@ -100,6 +103,81 @@ void main() {
       await tester.pumpAndSettle();
 
       verify(entryCubit.togglePrivate).called(1);
+    });
+
+    testWidgets('tap entry date', (WidgetTester tester) async {
+      when(entryCubit.togglePrivate).thenAnswer((_) async => true);
+
+      DateTime? modifiedDateTo;
+
+      when(() => entryCubit.updateFromTo(
+            dateFrom: testTextEntry.meta.dateFrom,
+            dateTo: any(named: 'dateTo'),
+          )).thenAnswer((Invocation i) async {
+        const dateTo = Symbol('dateTo');
+        modifiedDateTo = i.namedArguments[dateTo] as DateTime;
+        return true;
+      });
+
+      await tester.pumpWidget(
+        makeTestableWidgetWithScaffold(
+          BlocProvider<EntryCubit>.value(
+            value: entryCubit,
+            child: EntryDetailHeader(itemId: testTextEntry.meta.id),
+          ),
+        ),
+      );
+      await tester.pumpAndSettle();
+
+      final entryDateFromFinder =
+          find.text(df.format(testTextEntry.meta.dateFrom));
+      expect(entryDateFromFinder, findsOneWidget);
+
+      await tester.tap(entryDateFromFinder);
+      await tester.pumpAndSettle();
+
+      // entry date displayed in modal, too
+      expect(entryDateFromFinder, findsNWidgets(2));
+
+      // open and close dateTo selection
+      final entryDateToFinder = find.text(df.format(testTextEntry.meta.dateTo));
+      expect(entryDateToFinder, findsOneWidget);
+
+      await tester.tap(entryDateToFinder);
+      await tester.pumpAndSettle();
+
+      final doneButtonFinder = find.text('Done');
+      expect(doneButtonFinder, findsOneWidget);
+
+      await tester.tap(doneButtonFinder);
+      await tester.pumpAndSettle();
+
+      // open and close dateFrom selection
+      await tester.tap(entryDateFromFinder.last);
+      await tester.pumpAndSettle();
+
+      expect(doneButtonFinder, findsOneWidget);
+
+      await tester.tap(doneButtonFinder);
+      await tester.pumpAndSettle();
+
+      // set dateTo to now() and save
+      final nowButtonFinder = find.text('now');
+      expect(nowButtonFinder, findsOneWidget);
+
+      await tester.tap(nowButtonFinder);
+      await tester.pumpAndSettle();
+
+      final saveButtonFinder = find.text('SAVE');
+      expect(saveButtonFinder, findsOneWidget);
+
+      await tester.tap(saveButtonFinder);
+      await tester.pumpAndSettle();
+
+      verify(mockAppRouter.pop).called(1);
+
+      // updateFromTo called with recent dateTo after tapping now()
+      expect(modifiedDateTo?.difference(DateTime.now()).inSeconds, lessThan(2));
     });
   });
 }
